Refactor: Put the (linux) systemd and dbus based solutions to   separate modules.
