+++ wakepy/_linux/_systemd.py
@@ -1,0 +1,41 @@
+"""This module provides the systemd based solution for linux 
+
+See:
+https://www.man7.org/linux/man-pages/man1/systemctl.1.html
+"""
+
+import subprocess
+
+
+from wakepy.exceptions import NotSupportedError
+
+try:
+    subprocess.check_output(["pidof", "systemd"])
+except subprocess.CalledProcessError:
+    # if 'pidof' does not find a process it will return with non-zero exit
+    # status, check_output will raise subprocess.CalledProcessError
+    # See: https://github.com/np-8/wakepy/pull/3
+    raise NotSupportedError("sysmtemd not supported")
+
+
+def run_systemctl(command):
+    args = ["sleep.target", "suspend.target", "hibernate.target", "hybrid-sleep.target"]
+    subprocess.run(["systemctl", command, *args])
+
+
+def set_keepawake(keep_screen_awake=False):
+    """
+    Set the keep-awake. During keep-awake, the CPU is not allowed to go to
+    sleep automatically until the `unset_keepawake` is called.
+
+    Parameters
+    -----------
+    keep_screen_awake: bool
+        Currently unused as the screen will remain active as a byproduct of
+        preventing sleep.
+    """
+    run_systemctl("mask")
+
+
+def unset_keepawake():
+    run_systemctl("unmask")
